backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
httpx==0.27.2
python-dotenv==1.0.1
google-genai==0.3.0
librosa==0.10.2.post1
soundfile==0.12.1
elasticsearch==8.15.1
pydantic==2.9.2
pydantic-settings==2.5.2
